requirements.txt
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
Farama-Notifications==0.0.4
filelock==3.18.0
fonttools==4.57.0
frozendict==2.4.6
fsspec==2025.3.2
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.1.1
huggingface-hub==0.30.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
joblib>=1.2.0
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib>=3.4.0
mpmath==1.3.0
multitasking==0.0.11
networkx==3.4.2
numpy>=1.20.0
packaging==24.2
pandas>=1.3.0
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.7
plotly>=5.14.0
pluggy==1.5.0
pyparsing==3.2.3
pytest>=7.0.0
python-dateutil==2.9.0.post0
python-dotenv>=1.0.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests>=2.27.0
safetensors==0.5.3
scikit-learn>=1.2.0
scipy>=1.10.0
seaborn>=0.11.0
setuptools==75.8.0
six==1.17.0
soupsieve==2.6
stable_baselines3==2.6.0
statsmodels>=0.14.0
sympy==1.13.1
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.51.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
websocket-client>=1.3.0
wheel==0.45.1
yfinance>=0.1.70
quandl>=3.6.0
matplotlib-venn>=0.11.9
reportlab>=3.6.0
pdfkit>=1.0.0
WeasyPrint>=56.1
flask-mail>=0.9.1
Flask==2.2.3
pandas==1.5.3
numpy==1.24.2
matplotlib==3.7.1
scikit-learn==1.2.2
werkzeug==2.2.3
reportlab==3.6.12
python-dotenv==1.0.0
flask-mail==0.9.1
tensorflow==2.18.1
nltk==3.8.1
seaborn==0.12.2
plotly==5.14.0
transformers==4.28.1
shap==0.41.0

# Core dependencies
flask==2.3.3
pymongo==4.5.0
pyjwt==2.8.0
cryptography==41.0.4
bcrypt==4.0.1
ipaddress==1.0.23

# Web-related dependencies
flask-cors==4.0.0

# Utility libraries
python-dotenv==1.0.0
gunicorn==21.2.0
pytz==2023.3

# Testing tools
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# Environment helpers
python-decouple==3.8

# Database dependencies
redis==5.0.0
boto3==1.28.38
botocore==1.31.38

# Deployment tools
docker==6.1.3
kubernetes==28.1.0
google-cloud-storage==2.10.0
google-cloud-container==2.22.0

# Data fetching and analysis dependencies
yfinance>=0.2.32
pandas>=1.5.3
pandas_datareader>=0.10.0
lxml>=4.9.3
beautifulsoup4>=4.11.0
PyYAML>=6.0.1
requests>=2.31.0
numpy>=1.24.2

# AI and ML libraries
matplotlib>=3.7.1
scikit-learn>=1.2.2
tensorflow==2.18.1
nltk==3.8.1
seaborn==0.12.2
plotly==5.14.0
transformers==4.28.1
shap==0.41.0
stable_baselines3==2.6.0
gymnasium==1.1.1

# PDF Generation
reportlab==3.6.12
pdfkit>=1.0.0
WeasyPrint>=56.1

# Email
flask-mail==0.9.1
